third_party/blink/web_tests/crypto/subtle/hkdf/deriveBits-rfc5869-test-vectors-expected.txt

Tests deriveBits() using HKDF

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b] should be [3cb25f25faacd57a90434f64d0362f2a2d2d0a90cf1a5a4c5db02d56ecc4c5bf34007208d5b887185865] and was
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f404142434445464748494a4b4c4d4e4f] should be [b11e398dc80327a1c8e7f78c596a49344f012eda2d4efad8a050cc4c19afa97c59045a99cac7827271cb41c65e590e09da3275600c2f09b8367793a9aca3db71cc30c58179ec3e87c14c01d5c1f3434f1d87] and was
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b] should be [8da4e775a563c18f715f802a063c5a31b8a11f5c5ee1879ec3454e5f3c738d2d9d201395faa4b61a96c8] and was
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [0b0b0b0b0b0b0b0b0b0b0b] should be [085a01ea1b10f36933068b56efa5ad81a4f14b822f5b091568a9cdd4f155fda2c22e422478d305f3f896] and was
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f404142434445464748494a4b4c4d4e4f] should be [0bd770a74d1160f7c9f12cd5912a06ebff6adcae899d92191fe4305673ba2ffe8fa3f1a4e5ad79f3f334b3b202b2173c486ea37ce3d397ed034c7f9dfeb15c5e927336d0441f4c4300e2cff0d0900b52d3b4] and was
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b0b] should be [0ac1af7002b3d761d1e55298da9d0506b9ae52057220a306e07b6b87e8df21d0ea00033de03984d34918] and was
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HKDF"
PASS key.usages.join(',') is "deriveKey,deriveBits"
PASS: HKDF of [0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c0c] should be [2c91117204d745f3500d636a62f64f0ab3bae548aa53d423b0d1f27ebba6f5e5673a081d70cce7acfc48] and was
PASS successfullyParsed is true

TEST COMPLETE
